requirements.txt
flask==2.2
Flask-Cors==3.0.10
flask-jwt-extended==4.4.4
celery==5.2.5
redis==4.5.1
importlib-metadata==4.13.0
python-dotenv==1.0.0
setuptools

# docs
flasgger==0.9.5
marshmallow==3.19.0
apispec==6.3.0

# tests
coverage==7.2.5
requests==2.31.0
locust==2.15.1
